do bad

+++ src/day1.rs
@@ -1,7 +1,7 @@
 #[aoc(day1, part1)]
 pub fn part1(s: &str) -> i32 {
     let input = s;
-    input.len() as i32
+    input.len()
 }
 
 #[aoc(day1, part2)]
